Refactor types into strategy-pattern (ish) subclasses.
