+++ packages/random/package.js
@@ -6,7 +6,8 @@ Package.describe({
 Package.onUse(function (api) {
   api.use('ecmascript');
   api.export('Random');
-  api.mainModule('random.js');
+  api.mainModule('main_client.js', 'client');
+  api.mainModule('main_server.js', 'server');
 });
 
 Package.onTest(function (api) {
